src/webapp/requirements/base.txt
django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
djangorestframework==3.15.2
gunicorn
psycopg==3.2.4
